add detector
